Fix MPI_Bcast datatype for prob_num_auger

+++ nonthermal.c
@@ -3078,7 +3078,7 @@ void nt_MPI_Bcast(const int modelgridindex, const int root)
   MPI_Bcast(nt_solution[modelgridindex].fracdep_ionization_ion, includedions, MPI_DOUBLE, root, MPI_COMM_WORLD);
   MPI_Bcast(nt_solution[modelgridindex].eff_ionpot, includedions, MPI_FLOAT, root, MPI_COMM_WORLD);
 
-  MPI_Bcast(nt_solution[modelgridindex].prob_num_auger, includedions * (MAX_AUGER_ELECTRONS + 1), MPI_FLOAT, root, MPI_COMM_WORLD);
+  MPI_Bcast(nt_solution[modelgridindex].prob_num_auger, includedions * (MAX_AUGER_ELECTRONS + 1), MPI_DOUBLE, root, MPI_COMM_WORLD);
   MPI_Bcast(nt_solution[modelgridindex].ionenfrac_num_auger, includedions * (MAX_AUGER_ELECTRONS + 1), MPI_FLOAT, root, MPI_COMM_WORLD);
 
   // communicate NT excitations
